backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "gunicorn>=25.0.3",
    "jsonlines>=4.0.0",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
]
